Add business.
The add business form should be working now. The iPhone sends the name,
address, etc. to the server which does some processing then includes it on the
business list. When the add business modal popup goes away, the list is reloaded
with fresh data from the server.

Eventually, we'll have to handle pagination