Important Change: Names and Structure
Tags, Categories, BusinessTags, Sorts have ALL been changed to Topic and
BusinessTopic. Similarly, UserSubscription has been changed to UserTopic since
it related a user to a topic.

Topics are the attributes of a restaurant: characteristics that describe it.

Types are the buckets that describe the actual establishment (e.g. a coffee
shop).

Ratings are now a float value between 0 and 1 in the database. You can do
whatever conversion you want to an integer, but its easiest just to use floats
as the intermediate representation.

The schema is currently set up so that Topics form a tree (there is only one
parent of a Topic). I will change this in the next commit or two